requirements.txt
influxdb-client
python-dotenv
requests